RankingItemをranking-body.jsxに統合
