+++ resources/assets/jsx/components/ranking-body.jsx
@@ -1,10 +1,65 @@
 import React from 'react';
 import RankingTypes from '../../js/ranking-types';
-import RankingItem from './ranking-item.jsx';
 import Pagination from "./pagination.jsx";
 
 import { observer } from 'mobx-react'
 import rankingStore from '../../js/ranking-store';
+
+/**
+ * プレーヤーデータAPIの戻り値を表示できる形式に変換する
+ * @param json_data APIの戻り値
+ * @param rankingType string ランキングの種類
+ * @returns string
+ */
+const formatRankedData = (json_data, rankingType) => {
+    if (rankingType === "playtime") {
+        const {data} = json_data;
+        return `${data.hours}時間${data.minutes}分${data.seconds}秒`;
+    }
+
+    // 生のデータを下から3桁ずつカンマで区切っていく
+    const data = json_data["raw_data"];
+    let result = "";
+    for(let backward_index = 1; backward_index <= data.length; backward_index++) {
+        const index = data.length - backward_index;
+        result = data[index] + result;
+        if (index !== 0 && backward_index % 3 === 0) {
+            result = "," + result;
+        }
+    }
+
+    return result;
+};
+
+/**
+ * アバターのURLを取得する
+ * @param playerName プレーヤー名
+ */
+const getAvatarUrl = playerName => `https://mcapi.ca/avatar/${playerName}/60`;
+
+const RankingItem = observer(({rank: rankObject}) => {
+    const player = rankObject.player;
+
+    return (
+        <tr>
+            <th scope="row">
+                <span className="rank">
+                    {rankObject.rank}位
+                </span>
+            </th>
+            <td>
+                <img src={getAvatarUrl(player.name)}/>
+            </td>
+            <td>
+                {player.name}<br />
+                <span className="ranked-data">
+                        {RankingTypes.resolveRaw(rankObject.type)}：{formatRankedData(rankObject.data, rankObject.type)}
+                </span><br />
+                <span className="last_login">Last quit：{rankObject.lastquit}</span>
+            </td>
+        </tr>
+    );
+});
 
 const WrappedPagination = observer(({ store }) =>
     store.ranking !== undefined
@@ -13,7 +68,6 @@ const WrappedPagination = observer(({ store }) =>
                       onPageChange={ page => store.setPage(page) }/>
         : null
 );
-
 
 const RankingTable = observer(({ store }) => {
     // TODO 期間ランキングのAPIが実装され次第このブロックを消すこと
